refactor: rename RampRouter package to agno.models.ramp
Align the module path with the naming standard: agno.models.router ->
agno.models.ramp, with `ramp` as the vendor namespace and `router.py`
the Router model file (agno.models.ramp.router.RampRouter). The test and
cookbook directories are renamed to match. The provider-registry key is
now `ramp`, so the model string is `ramp:<id>`; `ramprouter:<id>` still
resolves via the display provider, and `router:<id>` no longer does.
