--- libs/agno/agno/models/router/__init__.py
@@ -1,3 +1,0 @@
-from agno.models.router.router import RampRouter
-
-__all__ = ["RampRouter"]
